+++ core/src/test/java/tech/guilhermekaua/spigotboot/core/test/utils/NumberUtilsTest.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package tech.guilhermekaua.spigotboot.test.utils;
+package tech.guilhermekaua.spigotboot.core.test.utils;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
